backend/requirements.txt
openai==1.12.0
reportlab==4.1.0
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0
cors==1.0.1
httpx==0.26.0 